rename quantizer_config.h
test=develop
