+++ paddle/fluid/inference/api/paddle_analysis_config.h
@@ -23,9 +23,9 @@
 
 // Here we include some header files with relative paths, for that in deploy,
 // the abstract path of this header file will be changed.
-#include "paddle_api.h"           // NOLINT
-#include "paddle_pass_builder.h"  // NOLINT
-#include "quantizer_config.h"     // NOLINT
+#include "paddle_api.h"               // NOLINT
+#include "paddle_pass_builder.h"      // NOLINT
+#include "paddle_quantizer_config.h"  // NOLINT
 
 namespace paddle {
