+++ paddle/fluid/inference/api/quantizer_config.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/inference/api/quantizer_config.h"
+#include "paddle/fluid/inference/api/paddle_quantizer_config.h"
 
 namespace paddle {
